+++ lib/epages/order.rb
@@ -6,7 +6,8 @@ module Epages
     include Epages::Utils
 
     DATE_ATTRS = %w(creation_date invoiced_on delivered_on pending_on archived_on dispatched_on viewed_on
-                    rejected_on closed_on paid_on returned_on)
+                    rejected_on closed_on paid_on returned_on partially_dispatched_on partially_paid_on
+                    partially_invoiced_on)
 
     ATTRS = %w(order_id order_number billing_address shipping_address customer_id locale currency_id tax_model
                grand_total total_before_tax line_item_container product_line_items shipping_price links internal_note